move quality flags into variables tab
