+++ lib/components/DialogBase/DialogBase.d.ts
@@ -10,7 +10,7 @@ declare namespace DialogBase {
         export const closeButtonProps: PropTypes.Requireable<{
             [x: string]: string | number | null | undefined;
         }>;
-        export const noPaper: PropTypes.Requireable<boolean>;
+        export const nopaper: PropTypes.Requireable<boolean>;
     }
     export namespace defaultProps {
         const open_1: boolean;
@@ -19,8 +19,8 @@ declare namespace DialogBase {
         export { toolbarChildren_1 as toolbarChildren };
         const closeButtonProps_1: {};
         export { closeButtonProps_1 as closeButtonProps };
-        const noPaper_1: boolean;
-        export { noPaper_1 as noPaper };
+        const nopaper_1: boolean;
+        export { nopaper_1 as nopaper };
     }
 }
 import PropTypes from "prop-types";
