+++ lib/components/DialogBase/DialogBase.js
@@ -81,8 +81,8 @@ var DialogBase = function DialogBase(props) {
       toolbarChildren = props.toolbarChildren,
       children = props.children,
       closeButtonProps = props.closeButtonProps,
-      noPaper = props.noPaper,
-      other = _objectWithoutProperties(props, ["open", "onClose", "title", "toolbarChildren", "children", "closeButtonProps", "noPaper"]);
+      nopaper = props.nopaper,
+      other = _objectWithoutProperties(props, ["open", "onClose", "title", "toolbarChildren", "children", "closeButtonProps", "nopaper"]);
 
   return /*#__PURE__*/_react.default.createElement(_Dialog.default, _extends({
     open: open,
@@ -106,7 +106,7 @@ var DialogBase = function DialogBase(props) {
     variant: "h5",
     className: classes.dialogTitle,
     noWrap: true
-  }, title), toolbarChildren)), noPaper ? _react.default.createElement("div", {
+  }, title), toolbarChildren)), nopaper ? _react.default.createElement("div", {
     className: classes.noPaper
   }, children) : _react.default.createElement(_Paper.default, {
     className: classes.contentPaper
@@ -120,13 +120,13 @@ DialogBase.propTypes = {
   toolbarChildren: _propTypes.default.oneOfType([_propTypes.default.arrayOf(_propTypes.default.oneOfType([_propTypes.default.node, _propTypes.default.string])), _propTypes.default.node, _propTypes.default.string]),
   children: _propTypes.default.oneOfType([_propTypes.default.arrayOf(_propTypes.default.oneOfType([_propTypes.default.node, _propTypes.default.string])), _propTypes.default.node, _propTypes.default.string]).isRequired,
   closeButtonProps: _propTypes.default.objectOf(_propTypes.default.oneOfType([_propTypes.default.string, _propTypes.default.number])),
-  noPaper: _propTypes.default.bool
+  nopaper: _propTypes.default.bool
 };
 DialogBase.defaultProps = {
   open: true,
   toolbarChildren: null,
   closeButtonProps: {},
-  noPaper: false
+  nopaper: false
 };
 var _default = DialogBase;
 exports.default = _default;
