+++ lib/components/TimeSeriesViewer/TimeSeriesViewerContainer.js
@@ -428,8 +428,11 @@ function TimeSeriesViewerContainer() {
           label = _TABS$tabId.label,
           ariaLabel = _TABS$tabId.ariaLabel,
           TabIcon = _TABS$tabId.Icon;
-      var style = {}; // if (tabId === TAB_IDS.SUMMARY && !belowMd) {
-      // style.borderTopLeftRadius = Theme.spacing(1); }
+      var style = {};
+
+      if (tabId === TAB_IDS.OPTIONS && !belowMd) {
+        style.borderBottomLeftRadius = _Theme.default.spacing(1);
+      }
 
       return _react.default.createElement(_Tab.default, {
         key: tabId,
