+++ lib/components/TimeSeriesViewer/TimeSeriesViewerGraph.js
@@ -555,6 +555,8 @@ function TimeSeriesViewerGraph() {
     axes.forEach(function (axis) {
       axesOption[axis.axis] = {
         independentTicks: true
+        /* valueRange: [low, high], */
+
       };
     });
     return axesOption;
