+++ lib/components/TimeSeriesViewer/TimeSeriesViewerOptions.js
@@ -13,12 +13,6 @@ var _styles = require("@material-ui/core/styles");
 
 var _Button = _interopRequireDefault(require("@material-ui/core/Button"));
 
-var _Checkbox = _interopRequireDefault(require("@material-ui/core/Checkbox"));
-
-var _FormGroup = _interopRequireDefault(require("@material-ui/core/FormGroup"));
-
-var _FormControlLabel = _interopRequireDefault(require("@material-ui/core/FormControlLabel"));
-
 var _Grid = _interopRequireDefault(require("@material-ui/core/Grid"));
 
 var _Slider = _interopRequireDefault(require("@material-ui/core/Slider"));
@@ -30,12 +24,6 @@ var _Skeleton = _interopRequireDefault(require("@material-ui/lab/Skeleton"));
 var _ToggleButton = _interopRequireDefault(require("@material-ui/lab/ToggleButton"));
 
 var _ToggleButtonGroup = _interopRequireDefault(require("@material-ui/lab/ToggleButtonGroup"));
-
-var _NotInterested = _interopRequireDefault(require("@material-ui/icons/NotInterested"));
-
-var _DoneAll = _interopRequireDefault(require("@material-ui/icons/DoneAll"));
-
-var _Clear = _interopRequireDefault(require("@material-ui/icons/Clear"));
 
 var _SwapHoriz = _interopRequireDefault(require("@material-ui/icons/SwapHoriz"));
 
@@ -59,25 +47,6 @@ function _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }
 
 var useStyles = (0, _styles.makeStyles)(function (theme) {
   return {
-    noneContainer: {
-      color: theme.palette.grey[400],
-      display: 'flex',
-      alignItems: 'flex-start'
-    },
-    noneIcon: {
-      color: theme.palette.grey[400],
-      margin: theme.spacing(0.375, 0.5, 0, 0),
-      fontSize: '1rem'
-    },
-    noneLabel: {
-      fontSize: '0.95rem'
-    },
-    qualityFlagsButtons: {
-      marginBottom: theme.spacing(2),
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'flex-start'
-    },
     smallButton: {
       fontSize: '0.8rem',
       padding: theme.spacing(0.125, 0.75),
@@ -172,14 +141,9 @@ var RollPeriodSlider = (0, _styles.withStyles)({
     }
   }
 })(_Slider.default);
-
-var ucWord = function ucWord(word) {
-  return "".concat(word.slice(0, 1).toUpperCase()).concat(word.slice(1).toLowerCase());
-};
 /**
    Roll Period Option
 */
-
 
 var sliderDefault = {
   value: 1
@@ -345,125 +309,6 @@ var YAxesOption = function YAxesOption() {
   }), "Swap Y Axes"));
 };
 /**
-   Quality Flags Option
-*/
-
-
-var QualityFlagsOption = function QualityFlagsOption() {
-  var classes = useStyles(_Theme.default);
-
-  var _TimeSeriesViewerCont5 = _TimeSeriesViewerContext.default.useTimeSeriesViewerState(),
-      _TimeSeriesViewerCont6 = _slicedToArray(_TimeSeriesViewerCont5, 2),
-      state = _TimeSeriesViewerCont6[0],
-      dispatch = _TimeSeriesViewerCont6[1];
-
-  var availableQualityFlags = state.availableQualityFlags;
-  var selectedQualityFlags = state.selection.qualityFlags;
-
-  var toggleFlag = function toggleFlag(qualityFlag) {
-    return function (event) {
-      dispatch({
-        type: 'selectToggleQualityFlag',
-        qualityFlag: qualityFlag,
-        selected: event.target.checked
-      });
-    };
-  };
-
-  if (!availableQualityFlags.size) {
-    return _react.default.createElement("div", {
-      className: classes.noneContainer
-    }, _react.default.createElement(_NotInterested.default, {
-      className: classes.noneIcon
-    }), _react.default.createElement(_Typography.default, {
-      variant: "body1",
-      className: classes.noneLabel
-    }, "No Quality Flags Available"));
-  }
-
-  var organizedQualityFlags = {
-    basic: [],
-    expanded: []
-  };
-  Array.from(availableQualityFlags).forEach(function (qf) {
-    if (!state.variables[qf]) {
-      return;
-    }
-
-    var downloadPkg = state.variables[qf].downloadPkg;
-    organizedQualityFlags[downloadPkg].push(qf);
-  });
-  organizedQualityFlags.basic.sort();
-  organizedQualityFlags.expanded.sort();
-  var downloadPkgs = ['basic', 'expanded'];
-  return _react.default.createElement(_react.default.Fragment, null, availableQualityFlags.size > 1 ? _react.default.createElement("div", {
-    className: classes.qualityFlagsButtons
-  }, _react.default.createElement(_Button.default, {
-    color: "primary",
-    variant: "outlined",
-    onClick: function onClick() {
-      dispatch({
-        type: 'selectNoneQualityFlags'
-      });
-    },
-    className: classes.smallButton,
-    style: {
-      marginRight: _Theme.default.spacing(2)
-    }
-  }, _react.default.createElement(_Clear.default, {
-    className: classes.smallButtonIcon
-  }), "Select None"), _react.default.createElement(_Button.default, {
-    color: "primary",
-    variant: "outlined",
-    onClick: function onClick() {
-      dispatch({
-        type: 'selectAllQualityFlags'
-      });
-    },
-    className: classes.smallButton
-  }, _react.default.createElement(_DoneAll.default, {
-    className: classes.smallButtonIcon
-  }), "Select All (".concat(availableQualityFlags.size, ")"))) : null, _react.default.createElement(_FormGroup.default, null, downloadPkgs.map(function (downloadPkg) {
-    return _react.default.createElement("div", {
-      key: downloadPkg
-    }, _react.default.createElement(_Typography.default, {
-      variant: "subtitle2"
-    }, ucWord(downloadPkg)), !organizedQualityFlags[downloadPkg].length ? _react.default.createElement(_Typography.default, {
-      variant: "body2",
-      className: classes.noneLabel
-    }, "No ".concat(downloadPkg, " quality flags available")) : _react.default.createElement(_react.default.Fragment, null, organizedQualityFlags[downloadPkg].map(function (qf) {
-      var checked = selectedQualityFlags.includes(qf);
-      var captionStyle = {
-        display: 'block',
-        color: _Theme.default.palette.grey[400]
-      };
-      return _react.default.createElement(_FormControlLabel.default, {
-        key: qf,
-        style: {
-          alignItems: 'flex-start',
-          marginBottom: _Theme.default.spacing(1)
-        },
-        control: _react.default.createElement(_Checkbox.default, {
-          value: qf,
-          color: "primary",
-          checked: checked,
-          onChange: toggleFlag(qf)
-        }),
-        label: _react.default.createElement("div", {
-          style: {
-            paddingTop: _Theme.default.spacing(0.5)
-          }
-        }, _react.default.createElement(_Typography.default, {
-          variant: "body2"
-        }, qf, _react.default.createElement(_Typography.default, {
-          variant: "caption",
-          style: captionStyle
-        }, state.variables[qf].description)))
-      });
-    })));
-  })));
-};
-/**
    Time Step Option
 */
 
@@ -471,10 +316,10 @@ var QualityFlagsOption = function QualityFlagsOption() {
 var TimeStepOption = function TimeStepOption() {
   var classes = useStyles(_Theme.default);
 
-  var _TimeSeriesViewerCont7 = _TimeSeriesViewerContext.default.useTimeSeriesViewerState(),
-      _TimeSeriesViewerCont8 = _slicedToArray(_TimeSeriesViewerCont7, 2),
-      state = _TimeSeriesViewerCont8[0],
-      dispatch = _TimeSeriesViewerCont8[1];
+  var _TimeSeriesViewerCont5 = _TimeSeriesViewerContext.default.useTimeSeriesViewerState(),
+      _TimeSeriesViewerCont6 = _slicedToArray(_TimeSeriesViewerCont5, 2),
+      state = _TimeSeriesViewerCont6[0],
+      dispatch = _TimeSeriesViewerCont6[1];
 
   var availableTimeSteps = state.availableTimeSteps;
   var selectedTimeStep = state.selection.timeStep;
@@ -519,11 +364,6 @@ var OPTIONS = {
     title: 'Roll Period',
     description: 'Set a rolling window to smooth out noisy data.',
     Component: RollPeriodOption
-  },
-  QUALITY_FLAGS: {
-    title: 'Quality Flags',
-    description: 'Enabling one or more quality flags will highlight regions on the chart to illustrate the results of data quality tests.',
-    Component: QualityFlagsOption
   },
   TIME_STEP: {
     title: 'Time Step',
@@ -570,8 +410,5 @@ function TimeSeriesViewerOptions() {
   }, renderOption('TIME_STEP')), _react.default.createElement(_Grid.default, {
     item: true,
     xs: 12
-  }, renderOption('ROLL_PERIOD')), _react.default.createElement(_Grid.default, {
-    item: true,
-    xs: 12
-  }, renderOption('QUALITY_FLAGS')));
+  }, renderOption('ROLL_PERIOD')));
 }
