+++ lib/components/TimeSeriesViewer/TimeSeriesViewerVariables.js
@@ -13,6 +13,14 @@ var _reactSelect = _interopRequireDefault(require("react-select"));
 
 var _styles = require("@material-ui/core/styles");
 
+var _Button = _interopRequireDefault(require("@material-ui/core/Button"));
+
+var _Checkbox = _interopRequireDefault(require("@material-ui/core/Checkbox"));
+
+var _FormControlLabel = _interopRequireDefault(require("@material-ui/core/FormControlLabel"));
+
+var _FormGroup = _interopRequireDefault(require("@material-ui/core/FormGroup"));
+
 var _IconButton = _interopRequireDefault(require("@material-ui/core/IconButton"));
 
 var _InputAdornment = _interopRequireDefault(require("@material-ui/core/InputAdornment"));
@@ -27,11 +35,15 @@ var _Paper = _interopRequireDefault(require("@material-ui/core/Paper"));
 
 var _Skeleton = _interopRequireDefault(require("@material-ui/lab/Skeleton"));
 
-var _MenuItem = _interopRequireDefault(require("@material-ui/core/MenuItem"));
-
 var _Clear = _interopRequireDefault(require("@material-ui/icons/Clear"));
 
+var _MenuItem = _interopRequireDefault(require("@material-ui/core/MenuItem"));
+
+var _NotInterested = _interopRequireDefault(require("@material-ui/icons/NotInterested"));
+
 var _Search = _interopRequireDefault(require("@material-ui/icons/Search"));
+
+var _DoneAll = _interopRequireDefault(require("@material-ui/icons/DoneAll"));
 
 var _Theme = _interopRequireDefault(require("../Theme/Theme"));
 
@@ -106,9 +118,48 @@ var useStyles = (0, _styles.makeStyles)(function (theme) {
     variablePaperContainer: {
       lineHeight: '5em',
       marginTop: theme.spacing(2.5)
+    },
+    noneContainer: {
+      color: theme.palette.grey[400],
+      display: 'flex',
+      alignItems: 'flex-start'
+    },
+    noneIcon: {
+      color: theme.palette.grey[400],
+      margin: theme.spacing(0.375, 0.5, 0, 0),
+      fontSize: '1rem'
+    },
+    noneLabel: {
+      fontSize: '0.95rem'
+    },
+    qualityFlagsContainer: {
+      marginTop: theme.spacing(2)
+    },
+    qualityFlagsHeading: {
+      fontWeight: 600,
+      marginBottom: _Theme.default.spacing(0.5)
+    },
+    qualityFlagsButtons: {
+      marginBottom: theme.spacing(2),
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'flex-start'
+    },
+    smallButton: {
+      fontSize: '0.8rem',
+      padding: theme.spacing(0.125, 0.75),
+      whiteSpace: 'nowrap'
+    },
+    smallButtonIcon: {
+      marginRight: theme.spacing(0.5),
+      fontSize: '0.8rem'
     }
   };
 });
+
+var ucWord = function ucWord(word) {
+  return "".concat(word.slice(0, 1).toUpperCase()).concat(word.slice(1).toLowerCase());
+};
 
 function inputComponent(_ref) {
   var inputRef = _ref.inputRef,
@@ -302,14 +353,132 @@ var selectStyles = {
     });
   }
 };
+/**
+   Quality Flags
+*/
 
-function TimeSeriesViewerVariables() {
+var QualityFlags = function QualityFlags() {
   var classes = useStyles(_Theme.default);
 
   var _TimeSeriesViewerCont = _TimeSeriesViewerContext.default.useTimeSeriesViewerState(),
       _TimeSeriesViewerCont2 = _slicedToArray(_TimeSeriesViewerCont, 2),
       state = _TimeSeriesViewerCont2[0],
       dispatch = _TimeSeriesViewerCont2[1];
+
+  var availableQualityFlags = state.availableQualityFlags;
+  var selectedQualityFlags = state.selection.qualityFlags;
+
+  var toggleFlag = function toggleFlag(qualityFlag) {
+    return function (event) {
+      dispatch({
+        type: 'selectToggleQualityFlag',
+        qualityFlag: qualityFlag,
+        selected: event.target.checked
+      });
+    };
+  };
+
+  if (!availableQualityFlags.size) {
+    return _react.default.createElement("div", {
+      className: classes.noneContainer
+    }, _react.default.createElement(_NotInterested.default, {
+      className: classes.noneIcon
+    }), _react.default.createElement(_Typography.default, {
+      variant: "body1",
+      className: classes.noneLabel
+    }, "No Quality Flags Available"));
+  }
+
+  var organizedQualityFlags = {
+    basic: [],
+    expanded: []
+  };
+  Array.from(availableQualityFlags).forEach(function (qf) {
+    if (!state.variables[qf]) {
+      return;
+    }
+
+    var downloadPkg = state.variables[qf].downloadPkg;
+    organizedQualityFlags[downloadPkg].push(qf);
+  });
+  organizedQualityFlags.basic.sort();
+  organizedQualityFlags.expanded.sort();
+  var downloadPkgs = ['basic', 'expanded'];
+  return _react.default.createElement(_react.default.Fragment, null, availableQualityFlags.size > 1 ? _react.default.createElement("div", {
+    className: classes.qualityFlagsButtons
+  }, _react.default.createElement(_Button.default, {
+    color: "primary",
+    variant: "outlined",
+    onClick: function onClick() {
+      dispatch({
+        type: 'selectNoneQualityFlags'
+      });
+    },
+    className: classes.smallButton,
+    style: {
+      marginRight: _Theme.default.spacing(2)
+    }
+  }, _react.default.createElement(_Clear.default, {
+    className: classes.smallButtonIcon
+  }), "Select None"), _react.default.createElement(_Button.default, {
+    color: "primary",
+    variant: "outlined",
+    onClick: function onClick() {
+      dispatch({
+        type: 'selectAllQualityFlags'
+      });
+    },
+    className: classes.smallButton
+  }, _react.default.createElement(_DoneAll.default, {
+    className: classes.smallButtonIcon
+  }), "Select All (".concat(availableQualityFlags.size, ")"))) : null, _react.default.createElement(_FormGroup.default, null, downloadPkgs.map(function (downloadPkg) {
+    return _react.default.createElement("div", {
+      key: downloadPkg
+    }, _react.default.createElement(_Typography.default, {
+      variant: "subtitle2"
+    }, ucWord(downloadPkg)), !organizedQualityFlags[downloadPkg].length ? _react.default.createElement(_Typography.default, {
+      variant: "body2",
+      className: classes.noneLabel
+    }, "No ".concat(downloadPkg, " quality flags available")) : _react.default.createElement(_react.default.Fragment, null, organizedQualityFlags[downloadPkg].map(function (qf) {
+      var checked = selectedQualityFlags.includes(qf);
+      var captionStyle = {
+        display: 'block',
+        color: _Theme.default.palette.grey[400]
+      };
+      return _react.default.createElement(_FormControlLabel.default, {
+        key: qf,
+        style: {
+          alignItems: 'flex-start',
+          marginBottom: _Theme.default.spacing(1)
+        },
+        control: _react.default.createElement(_Checkbox.default, {
+          value: qf,
+          color: "primary",
+          checked: checked,
+          onChange: toggleFlag(qf)
+        }),
+        label: _react.default.createElement("div", {
+          style: {
+            paddingTop: _Theme.default.spacing(0.5)
+          }
+        }, _react.default.createElement(_Typography.default, {
+          variant: "body2"
+        }, qf, _react.default.createElement(_Typography.default, {
+          variant: "caption",
+          style: captionStyle
+        }, state.variables[qf].description)))
+      });
+    })));
+  })));
+};
+
+function TimeSeriesViewerVariables() {
+  var classes = useStyles(_Theme.default);
+
+  var _TimeSeriesViewerCont3 = _TimeSeriesViewerContext.default.useTimeSeriesViewerState(),
+      _TimeSeriesViewerCont4 = _slicedToArray(_TimeSeriesViewerCont3, 2),
+      state = _TimeSeriesViewerCont4[0],
+      dispatch = _TimeSeriesViewerCont4[1];
 
   var selectedVariables = state.selection.variables.map(function (variable) {
     return _extends({}, state.variables[variable], {
@@ -421,5 +590,20 @@ function TimeSeriesViewerVariables() {
       className: classes.optionSubtitle,
       gutterBottom: true
     }, description)));
-  })));
+  })), _react.default.createElement("div", {
+    className: classes.qualityFlagsContainer
+  }, _react.default.createElement(_Typography.default, {
+    variant: "subtitle1",
+    className: classes.qualityFlagsHeading
+  }, "Quality Flags"), _react.default.createElement(_Typography.default, {
+    variant: "caption",
+    style: {
+      color: _Theme.default.palette.grey[400]
+    }
+  }, "Enabling one or more quality flags will highlight regions on the chart to illustrate the results of data quality tests."), _react.default.createElement("div", {
+    style: {
+      width: '100%',
+      marginTop: _Theme.default.spacing(1)
+    }
+  }, _react.default.createElement(QualityFlags, null))));
 }
